refactor to move operations to another file
