+++ src/main/kotlin/tech/challenge/fastfood/fastfood/infra/adapters/controllers/ProductController.kt
@@ -1,0 +1,9 @@
+package tech.challenge.fastfood.fastfood.infra.adapters.controllers
+
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+@RequestMapping("/api/customers")
+class ProductController {
+}
