+++ src/main/kotlin/tech/challenge/fastfood/fastfood/infra/adapters/controllers/operation/ProductOperation.kt
@@ -1,0 +1,4 @@
+package tech.challenge.fastfood.fastfood.infra.adapters.controllers.operation
+
+interface ProductOperation {
+}
